+++ CHANGELOG.md
@@ -4,7 +4,14 @@
 
 -   **terra-info** has been removed.
 
-# 12.0.1 (unreleased)
+# 12.1.0 (12.10.2021)
+
+### Features
+
+-   **terra-filter-toolbar** added an optional input to search through the table's datasource, based on the filter parameters.
+-   **router** - we're now supporting material icons on routes.
+
+# 12.0.1 (06.10.2021)
 
 ### Bug fixes